+++ src/soufix/fight/ia/type/IA83.java
@@ -101,7 +101,7 @@ public class IA83 extends AbstractNeedSpell
         }
       }
 
-      long percentPdv=(this.fighter.getPdv()*100)/this.fighter.getPdvMax();
+      double percentPdv=(this.fighter.getPdv()*100)/this.fighter.getPdvMax();
 
       if(this.fighter.getCurPa(this.fight)>0&&!action&&percentPdv<50&&percentPdv<95)
       {
